requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.3.0
autobahn==23.6.2
Automat==25.4.16
babel==2.17.0
backports.asyncio.runner==1.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.2
ddt==1.7.2
distro==1.9.0
docstring_parser==0.17.0
docutils==0.21.2
exceptiongroup==1.3.0
filetype==1.2.0
GeoAlchemy2==0.17.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
json-cfg-rw==0.5.0
kwonly-args==1.0.10
MarkupSafe==3.0.3
packaging==25.0
pem==23.1.0
pluggy==1.6.0
psutil==5.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytmpdir==1.2.6
pytz==2025.2
reactivex==4.0.4
requests==2.32.5
Rx==1.6.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==7.4.7
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.43
tomli==2.2.1
tqdm==4.67.1
treelib==1.8.0
Twisted==22.10.0
txaio==23.6.1
txhttputil==1.2.8
txwebsocket==1.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
vcversioner==2.16.0.0
vortexpy==3.4.3
zipp==3.23.0
zope.interface==8.0.1
